requirements.txt
Flask==2.2.2
flask_swagger_ui==4.11.1
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.3.0
pymongo==4.3.3
python-decouple==3.6
protobuf==3.20.3
scipy==1.7.3
tensorflow==1.13.1
keras==2.0.8
torch==1.13.1
torchvision
h5py==2.10.0
waitress==2.1.2
